refactor(1013): remove SelfExportUtils facade (SC-001, position 7)
Extract SelfExportUtils (40 LOC, 2 methods) from MistHelper.py to
src/export/self_export_utils.py. New file scores 10.00/10 pylint.

- Class methods use lazy `importlib.import_module("MistHelper")` to
  reach live apisession + DataExporter/DataProcessingUtils/TimeUtils
  without circular import.
- Top-level re-export preserves the direct reference at the menu
  tuple (MistHelper.py:18131 -- option 23).
- MistHelper.py pylint stable at 8.66/10 (no regression from P6
  baseline); +1 E0401 and +1 C0413 unavoidable per-extraction cost.
- Systematic test: 66/66 passed.
